Minor fixes

+++ src/common/GeneralFunctions.cpp
@@ -246,13 +246,13 @@ time_t getFileModificationTime(const char *filename)
 
 
 
-size_t getFileSize(const char *filename)
+long long getFileSize(const char *filename)
 {
   try
   {
     struct stat buf;
     if (stat(filename,&buf) == 0)
-      return buf.st_size;
+      return (long long)buf.st_size;
 
     return -1;
   }
